+++ src/zimscraperlib/rewriting/js.py
@@ -154,13 +154,15 @@ def create_js_rules() -> list[TransformationRule]:
     return [
         # rewriting `eval(...)` - invocation
         (re.compile(r"(?:^|\s)\beval\s*\("), replace_prefix_from(eval_str, "eval")),
+        (re.compile(r"\([\w]+,\s*eval\)\("), m2str(lambda _: f" {eval_str}")),
         # rewriting `x = eval` - no invocation
         (re.compile(r"[=]\s*\beval\b(?![(:.$])"), replace("eval", "self.eval")),
+        (re.compile(r"var\s+self"), replace("var", "let")),
         # rewriting `.postMessage` -> `__WB_pmw(self).postMessage`
         (re.compile(r"\.postMessage\b\("), add_prefix(".__WB_pmw(self)")),
         # rewriting `location = ` to custom expression `(...).href =` assignement
         (
-            re.compile(r"[^$.]?\s?\blocation\b\s*[=]\s*(?![\s\d=])"),
+            re.compile(r"(?:^|[^$.+*/%^-])\s?\blocation\b\s*[=]\s*(?![\s\d=])"),
             add_suffix_non_prop(check_loc),
         ),
         # rewriting `return this`
@@ -186,6 +188,7 @@ def create_js_rules() -> list[TransformationRule]:
         # As the rule will match first, it will prevent next rule matching `import` to
         # be apply to `async import`.
         (re.compile(r"async\s+import\s*\("), m2str(lambda x: x)),
+        (re.compile(r"[^$.]\bimport\s*\([^)]*\)\s*\{"), m2str(lambda x: x)),
         # esm dynamic import, if found, mark as module
         (
             re.compile(r"[^$.]\bimport\s*\("),